# backend/openapi.yaml
openapi: 3.0.3
info:
  title: TCG Collection Manager API
  description: REST API for managing trading card game collections, including games, sets, cards, user collections, decks, and booster packs.
  version: 0.2.0
servers:
  - url: http://localhost:8000
    description: Local development server
components:
  securitySchemes:
    AdminApiKey:
      type: http
      scheme: bearer
      description: Admin API key for protected endpoints

    SessionCookie:
      type: apiKey
      in: cookie
      name: session_id
      description: Session cookie for authenticated user endpoints
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
          example: "Game not found"
...
    CardQuantityUpdate:
      type: object
      required:
        - card_id
        - quantity
      properties:
        card_id:
          type: integer
          format: int64
          example: 1
        quantity:
          type: integer
          format: int64
          description: The new quantity (0 to remove from collection)
          example: 3

    # Deck schemas
    CreateDeckRequest:
      type: object
      required:
        - collection_id
        - name
      properties:
        collection_id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "Electric Theme Deck"
    DeckResponse:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          format: int64
          example: 1
    UpdateDeckRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: "My Renamed Deck"
    DeckListItem:
      type: object
      required:
        - id
        - collection_id
        - name
        - created_at
        - collection_name
        - game_name
        - card_count
      properties:
        id:
          type: integer
          format: int64
          example: 1
        collection_id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "Electric Theme Deck"
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        collection_name:
          type: string
          example: "My Pokemon Collection"
        game_name:
          type: string
          example: "Pokemon TCG"
        game_image_url:
          type: string
          format: uri
          nullable: true
          example: "https://example.com/pokemon-logo.png"
        card_count:
          type: integer
          format: int64
          example: 60
    DeckCardItem:
      type: object
      required:
        - id
        - name
        - collector_number
        - attributes
        - set_id
        - set_name
        - quantity
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "Pikachu"
        collector_number:
          type: string
          example: "025"
        image_url:
          type: string
          format: uri
          nullable: true
          example: "https://example.com/pikachu.png"
        attributes:
          type: object
          additionalProperties:
            type: string
          example:
            rarity: "common"
            type: "electric"
        set_id:
          type: integer
          format: int64
          example: 1
        set_name:
          type: string
          example: "Shrouded Fable"
        quantity:
          type: integer
          format: int64
          example: 4

    # Booster schemas
...
paths:
...
  /api/decks:
    get:
      summary: List user's decks
      description: Get all decks owned by the authenticated user
      operationId: listDecks
      tags:
        - Decks
      security:
        - SessionCookie: []
      responses:
        '200':
          description: List of decks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeckListItem'
        '401':
          description: Unauthorized - no valid session
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create a deck
      description: Create a new deck within a collection
      operationId: createDeck
      tags:
        - Decks
      security:
        - SessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDeckRequest'
      responses:
        '200':
          description: Deck created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeckResponse'
        '401':
          description: Unauthorized - no valid session
        '403':
          description: Forbidden - collection not owned by user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Access denied"
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Collection not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/decks/{deck_id}:
    get:
      summary: Get deck details
      description: Get details of a specific deck
      operationId: getDeck
      tags:
        - Decks
      security:
        - SessionCookie: []
      parameters:
        - name: deck_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The deck ID
          example: 1
      responses:
        '200':
          description: Deck details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeckListItem'
        '401':
          description: Unauthorized - no valid session
        '403':
          description: Forbidden - deck not owned by user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Access denied"
        '404':
          description: Deck not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Deck not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      summary: Update deck
      description: Update the name of a deck
      operationId: updateDeck
      tags:
        - Decks
      security:
        - SessionCookie: []
      parameters:
        - name: deck_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The deck ID
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDeckRequest'
      responses:
        '204':
          description: Deck updated successfully
        '401':
          description: Unauthorized - no valid session
        '403':
          description: Forbidden - deck not owned by user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Access denied"
        '404':
          description: Deck not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Deck not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete deck
      description: Delete a deck and all associated cards
      operationId: deleteDeck
      tags:
        - Decks
      security:
        - SessionCookie: []
      parameters:
        - name: deck_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The deck ID
          example: 1
      responses:
        '204':
          description: Deck deleted successfully
        '401':
          description: Unauthorized - no valid session
        '403':
          description: Forbidden - deck not owned by user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Access denied"
        '404':
          description: Deck not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Deck not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/decks/{deck_id}/cards:
    get:
      summary: List cards in deck
      description: Get all cards in a specific deck with quantities
      operationId: listDeckCards
      tags:
        - Decks
      security:
        - SessionCookie: []
      parameters:
        - name: deck_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The deck ID
          example: 1
      responses:
        '200':
          description: List of cards in deck
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeckCardItem'
        '401':
          description: Unauthorized - no valid session
        '403':
          description: Forbidden - deck not owned by user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Access denied"
        '404':
          description: Deck not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Deck not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      summary: Update deck card quantities
      description: Update the quantities of cards in a deck. Cards must exist in the parent collection. Quantity cannot exceed collection quantity.
      operationId: updateDeckCards
      tags:
        - Decks
      security:
        - SessionCookie: []
      parameters:
        - name: deck_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The deck ID
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CardQuantityUpdate'
            example:
              - card_id: 1
                quantity: 4
              - card_id: 5
                quantity: 2
      responses:
        '204':
          description: Deck cards updated successfully
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                cardNotInCollection:
                  value:
                    error: "Card not in collection"
                insufficientQuantity:
                  value:
                    error: "Quantity exceeds collection"
        '401':
          description: Unauthorized - no valid session
        '403':
          description: Forbidden - deck not owned by user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Access denied"
        '404':
          description: Deck not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Deck not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...